requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses==0.6
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.3
multidict==5.0.2
numba==0.51.2
numpy==1.19.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.7.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.6.3
